+++ apps/client/app/login/page.tsx
@@ -1,4 +1,4 @@
-import NeedLogIn from '@/containers/needLogIn';
+import NeedLogIn from '@/containers/needLogin';
 
 export default async function LogIn(): Promise<React.ReactNode> {
   return <NeedLogIn />;
